issue-67: Create storage_lsr sub-directory under module_utils and move size.py there
https://github.com/linux-system-roles/storage/issues/67
Have to rename a couple of imports to handle this case.
Another issue is that python 2.x does not find the module unless
it has an `__init__.py` in the `storage_lsr/` sub-directory.
